+++ src/_log.jl
@@ -52,10 +52,8 @@ function log!(logger::Logger, state::SysState)
     logger.alpha4_vec[i] = state.alpha4
     logger.CL2_vec[i] = state.CL2
     logger.CD2_vec[i] = state.CD2
-    logger.aero_force_b_vec[i] .= state.aero_force_b
-    logger.aero_moment_b_vec[i] .= state.aero_moment_b
-    logger.tether_induced_force_vec[i] .= state.tether_induced_force
-    logger.tether_induced_moment_vec[i] .= state.tether_induced_moment
+    logger.aero_force_KA_vec[i] .= state.aero_force_KA
+    logger.aero_moment_KA_vec[i] .= state.aero_moment_KA
     logger.twist_angles_vec[i] .= state.twist_angles
     logger.vel_kite_vec[i] .= state.vel_kite
     logger.acc_vec[i] = state.acc